# doc/API_Documentation.openapi.yaml
openapi: 3.1.0
...
info:
  title: Core User Management API
  version: 1.0.0-draft
  summary: Reusable OpenAPI 3.1 contract for account lifecycle, authentication, user profile management, MFA, passkeys, admin user operations, and privacy preferences.
  description: |
    Canonical OpenAPI 3.1 contract for the reusable core user-management layer in this repository.

    This spec intentionally excludes the PriceTracker business-domain APIs such as catalog, shops, prices,
    purchases, comparisons, watchlists, alerts, and moderation of price submissions.

    The retained surface is limited to reusable platform concerns:

    - registration, login, logout, refresh, password reset, and MFA completion
    - current-user profile, avatar attachment, and contact management
    - passkeys, TOTP MFA, recovery codes, and security summary
    - account-owned private file handling needed by the core, such as avatars or export packages
    - admin user management and system settings
    - legal document discovery, privacy consent capture, and cookie preferences

    Contract conventions:

    - IDs use `format: uuid`
    - timestamps use `format: date-time` in UTC
    - default currency uses ISO 4217 alpha-3 codes
    - response bodies use `camelCase`
    - successful responses use `{ data, meta }`
    - error responses use `{ error }` and include `urgencyLevel` on a 1-9 scale
    - request DTOs reject unknown fields unless explicitly stated otherwise
servers:
...
    description: Production
...
paths:
...
  /me:
    get:
      tags:
        - Users
      operationId: getMe
      summary: Get the current user profile
      description: Return the authenticated user's profile, roles, scopes, and current security summary.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags:
        - Users
      operationId: updateMe
      summary: Update the current user profile
      description: Update editable profile and presentation settings. Clients may send `If-Match` for optimistic concurrency.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
  /me/avatar:
    post:
      tags:
        - Users
      operationId: setAvatar
      summary: Attach a caller-owned file as the active avatar
      description: The referenced file must already belong to the caller and be valid for avatar use.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvatarUpdateRequest'
      responses:
        '200':
          description: Updated profile with active avatar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Users
      operationId: removeAvatar
      summary: Remove the active avatar
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Updated profile without an avatar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/account/deactivate:
    post:
      tags:
        - Users
      operationId: deactivateOwnAccount
      summary: Deactivate the current account
      description: Deactivate the signed-in account. This is intended for self-service deactivation and can be paired with a later deletion or privacy request workflow.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDeactivateRequest'
      responses:
        '200':
          description: Account deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /me/sessions:
    get:
      tags:
        - IdentitySecurity
      operationId: listOwnSessions
      summary: List active sessions for the current account
      description: Return active and recently active sessions with device, IP, and last-seen metadata.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/sessions/revoke-all:
    post:
      tags:
        - IdentitySecurity
      operationId: revokeAllOwnSessions
      summary: Revoke all sessions for the current account
      description: Revoke all sessions or all sessions except the current one, depending on the requested scope.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionBulkRevokeRequest'
      responses:
        '200':
          description: Sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/sessions/{sessionId}:
    delete:
      tags:
        - IdentitySecurity
      operationId: revokeOwnSession
      summary: Revoke one session belonging to the current account
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      responses:
        '204':
          description: Session revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/security:
    get:
      tags:
        - IdentitySecurity
      operationId: getMySecuritySummary
      summary: Get current security status
      description: Return summary state needed by an account security page, such as MFA enablement and passkey count.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Security summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSecuritySummaryEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/security/events:
    get:
      tags:
        - IdentitySecurity
      operationId: listOwnSecurityEvents
      summary: List recent security events for the current account
      description: Return recent login, MFA, password, and session events visible to the current account.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Security event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityEventListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/security/reports:
    post:
      tags:
        - IdentitySecurity
      operationId: createOwnSecurityReport
      summary: Report suspicious login or account activity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityReportCreateRequest'
      responses:
        '201':
          description: Security report created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/passkeys:
    get:
      tags:
        - IdentitySecurity
      operationId: listPasskeys
      summary: List registered passkeys
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Passkey list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasskeyListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/passkeys/registration/options:
    post:
      tags:
        - IdentitySecurity
      operationId: createPasskeyRegistrationOptions
      summary: Create a WebAuthn registration challenge
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasskeyRegistrationOptionsRequest'
      responses:
        '200':
          description: WebAuthn registration options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasskeyRegistrationOptionsEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/passkeys/registration/verify:
    post:
      tags:
        - IdentitySecurity
      operationId: verifyPasskeyRegistration
      summary: Verify a WebAuthn registration result
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasskeyRegistrationVerifyRequest'
      responses:
        '201':
          description: Passkey registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasskeyEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/passkeys/{passkeyId}:
    delete:
      tags:
        - IdentitySecurity
      operationId: deletePasskey
      summary: Delete one registered passkey
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PasskeyIdPath'
      responses:
        '204':
          description: Passkey deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/mfa/totp/setup:
    post:
      tags:
        - IdentitySecurity
      operationId: createTotpSetup
      summary: Create a TOTP provisioning payload
      security:
        - bearerAuth: []
      responses:
        '200':
          description: TOTP secret and QR payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotpSetupEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/mfa/totp/enable:
    post:
      tags:
        - IdentitySecurity
      operationId: enableTotp
      summary: Verify a TOTP code and enable MFA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TotpEnableRequest'
      responses:
        '200':
          description: Updated security summary after enabling TOTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSecuritySummaryEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/mfa/totp/disable:
    post:
      tags:
        - IdentitySecurity
      operationId: disableTotp
      summary: Disable TOTP-based MFA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationCodeRequest'
      responses:
        '200':
          description: Updated security summary after disabling TOTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSecuritySummaryEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/mfa/recovery-codes/rotate:
    post:
      tags:
        - IdentitySecurity
      operationId: rotateRecoveryCodes
      summary: Replace the current recovery code set
      security:
        - bearerAuth: []
      responses:
        '200':
          description: New recovery codes generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryCodeListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/emails:
    get:
      tags:
        - IdentitySecurity
      operationId: listEmails
      summary: List the caller email addresses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Email address list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailAddressListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - IdentitySecurity
      operationId: createEmail
      summary: Add a secondary email address
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailAddressCreateRequest'
      responses:
        '201':
          description: Secondary email created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailAddressEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /me/emails/{emailId}:
    delete:
      tags:
        - IdentitySecurity
      operationId: deleteEmail
      summary: Delete one email address
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmailIdPath'
      responses:
        '204':
          description: Email deleted
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/emails/{emailId}/verify:
    post:
      tags:
        - IdentitySecurity
      operationId: verifyEmail
      summary: Verify one email address
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmailIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationCodeRequest'
      responses:
        '200':
          description: Verified email address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailAddressEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/emails/{emailId}/make-primary:
    post:
      tags:
        - IdentitySecurity
      operationId: makeEmailPrimary
      summary: Make one verified email primary
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmailIdPath'
      responses:
        '200':
          description: Updated primary email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailAddressEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/emails/{emailId}/resend-verification:
    post:
      tags:
        - IdentitySecurity
      operationId: resendEmailVerification
      summary: Resend a verification challenge for one email address
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmailIdPath'
      responses:
        '202':
          description: Verification challenge accepted for delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /me/phones/{phoneId}:
    delete:
      tags:
        - IdentitySecurity
      operationId: deletePhone
      summary: Delete one phone number
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PhoneIdPath'
      responses:
        '204':
          description: Phone number deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: false
      schema:
        type: string
      example: W/"rev-3"
...